+++ src/main/java/com/example/demo/shareItem/ShareItem.java
@@ -1,19 +1,33 @@
 package com.example.demo.shareItem;
 
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.Date;
+import java.util.List;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.OrderBy;
+import javax.persistence.SequenceGenerator;
+import javax.persistence.Table;
+
 import com.example.demo.shareDataDaily.ShareDataDaily;
 import com.example.demo.shareDataWeekly.ShareDataWeekly;
 import com.example.demo.utils.CurrencyEnum;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.*;
-import org.hibernate.annotations.Fetch;
-import org.hibernate.annotations.FetchMode;
 
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.Currency;
-import java.util.Date;
-import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 
 
 
@@ -45,7 +59,7 @@ public class ShareItem {
     private String symbol;
     @Column(nullable = false)
     @NonNull
-    private Double price;
+    private BigDecimal price;
 
     @Column(nullable = false)
     @NonNull
